+++ packages/language-core/lib/codegen/template/slotOutlet.ts
@@ -4,9 +4,10 @@ import { createVBindShorthandInlayHintInfo } from '../inlayHints';
 import { endOfLine, newLine, wrapWith } from '../utils';
 import type { TemplateCodegenContext } from './context';
 import { generateElementChildren } from './elementChildren';
-import { generateElementProps } from './elementProps';
+import { generateElementProps, generatePropExp } from './elementProps';
 import type { TemplateCodegenOptions } from './index';
 import { generateInterpolation } from './interpolation';
+import { generatePropertyAccess } from './propertyAccess';
 
 export function* generateSlotOutlet(
 	options: TemplateCodegenOptions,
@@ -14,7 +15,7 @@ export function* generateSlotOutlet(
 	node: CompilerDOM.SlotOutletNode
 ): Generator<Code> {
 	const startTagOffset = node.loc.start.offset + options.template.content.slice(node.loc.start.offset).indexOf(node.tag);
-	const varSlot = ctx.getInternalVariable();
+	const propsVar = ctx.getInternalVariable();
 	const nameProp = node.props.find(prop => {
 		if (prop.type === CompilerDOM.NodeTypes.ATTRIBUTE) {
 			return prop.name === 'name';
@@ -30,23 +31,59 @@ export function* generateSlotOutlet(
 
 	if (options.hasDefineSlots) {
 		yield `__VLS_normalizeSlot(`;
-		yield* wrapWith(
-			node.loc.start.offset,
-			node.loc.end.offset,
-			ctx.codeFeatures.verification,
-			`${options.slotsAssignName ?? '__VLS_slots'}[`,
-			...wrapWith(
+		if (nameProp) {
+			let codes: Generator<Code> | Code[];
+			if (nameProp.type === CompilerDOM.NodeTypes.ATTRIBUTE && nameProp.value) {
+				let { source, start: { offset } } = nameProp.value.loc;
+				if (source.startsWith('"') || source.startsWith("'")) {
+					source = source.slice(1, -1);
+					offset++;
+				}
+				codes = generatePropertyAccess(
+					options,
+					ctx,
+					source,
+					offset,
+					ctx.codeFeatures.navigationAndVerification
+				);
+			}
+			else if (
+				nameProp.type === CompilerDOM.NodeTypes.DIRECTIVE
+				&& nameProp.exp?.type === CompilerDOM.NodeTypes.SIMPLE_EXPRESSION
+			) {
+				codes = [
+					`[`,
+					...generatePropExp(
+						options,
+						ctx,
+						nameProp,
+						nameProp.exp,
+						ctx.codeFeatures.all,
+						true
+					),
+					`]`
+				];
+			}
+			else {
+				codes = [`['default']`];
+			}
+
+			yield* wrapWith(
+				nameProp.loc.start.offset,
+				nameProp.loc.end.offset,
+				ctx.codeFeatures.verification,
+				`${options.slotsAssignName ?? '__VLS_slots'}`,
+				...codes
+			);
+		}
+		else {
+			yield* wrapWith(
 				node.loc.start.offset,
 				node.loc.end.offset,
 				ctx.codeFeatures.verification,
-				nameProp?.type === CompilerDOM.NodeTypes.ATTRIBUTE && nameProp.value
-					? `'${nameProp.value.content}'`
-					: nameProp?.type === CompilerDOM.NodeTypes.DIRECTIVE && nameProp.exp?.type === CompilerDOM.NodeTypes.SIMPLE_EXPRESSION
-						? nameProp.exp.content
-						: `('default' as const)`
-			),
-			`]`
-		);
+				`${options.slotsAssignName ?? '__VLS_slots'}['default']`
+			);
+		}
 		yield `)?.(`;
 		yield* wrapWith(
 			startTagOffset,
@@ -66,7 +103,7 @@ export function* generateSlotOutlet(
 		yield `)${endOfLine}`;
 	}
 	else {
-		yield `var ${varSlot} = {${newLine}`;
+		yield `var ${propsVar} = {${newLine}`;
 		yield* generateElementProps(
 			options,
 			ctx,
@@ -83,10 +120,10 @@ export function* generateSlotOutlet(
 		) {
 			ctx.slots.push({
 				name: nameProp.value.content,
-				loc: nameProp.loc.start.offset + nameProp.loc.source.indexOf(nameProp.value.content, nameProp.name.length),
+				offset: nameProp.loc.start.offset + nameProp.loc.source.indexOf(nameProp.value.content, nameProp.name.length),
 				tagRange: [startTagOffset, startTagOffset + node.tag.length],
-				varName: varSlot,
 				nodeLoc: node.loc,
+				propsVar,
 			});
 		}
 		else if (
@@ -97,8 +134,8 @@ export function* generateSlotOutlet(
 			if (isShortHand) {
 				ctx.inlayHints.push(createVBindShorthandInlayHintInfo(nameProp.exp.loc, 'name'));
 			}
-			const slotExpVar = ctx.getInternalVariable();
-			yield `var ${slotExpVar} = `;
+			const expVar = ctx.getInternalVariable();
+			yield `var ${expVar} = __VLS_tryAsConstant(`;
 			yield* generateInterpolation(
 				options,
 				ctx,
@@ -106,22 +143,20 @@ export function* generateSlotOutlet(
 				ctx.codeFeatures.all,
 				nameProp.exp.content,
 				nameProp.exp.loc.start.offset,
-				nameProp.exp,
-				'(',
-				')'
+				nameProp.exp
 			);
-			yield ` as const${endOfLine}`;
+			yield `)${endOfLine}`;
 			ctx.dynamicSlots.push({
-				expVar: slotExpVar,
-				varName: varSlot,
+				expVar,
+				propsVar,
 			});
 		}
 		else {
 			ctx.slots.push({
 				name: 'default',
 				tagRange: [startTagOffset, startTagOffset + node.tag.length],
-				varName: varSlot,
 				nodeLoc: node.loc,
+				propsVar,
 			});
 		}
 	}