+++ packages/language-core/lib/parsers/scriptSetupRanges.ts
@@ -27,9 +27,14 @@ type DefineProps = CallExpressionRange & {
 	destructured?: Map<string, ts.Expression | undefined>;
 	destructuredRest?: string;
 	statement: TextRange;
+	// used by component-meta
+	argNode?: ts.Expression;
 };
 
-type WithDefaults = Pick<CallExpressionRange, 'callExp' | 'exp' | 'arg'>;
+type WithDefaults = Omit<CallExpressionRange, 'typeArg'> & {
+	// used by component-meta
+	argNode?: ts.Expression;
+};
 
 type DefineEmits = CallExpressionRange & {
 	name?: string;
@@ -39,7 +44,6 @@ type DefineEmits = CallExpressionRange & {
 
 type DefineSlots = CallExpressionRange & {
 	name?: string;
-	isObjectBindingPattern?: boolean;
 	statement: TextRange;
 };
 
@@ -281,7 +285,8 @@ export function parseScriptSetupRanges(
 			else if (vueCompilerOptions.macros.defineProps.includes(callText)) {
 				defineProps = {
 					...parseCallExpression(node),
-					statement: getStatementRange(ts, parents, node, ast)
+					statement: getStatementRange(ts, parents, node, ast),
+					argNode: node.arguments[0]
 				};
 				if (ts.isVariableDeclaration(parent)) {
 					if (ts.isObjectBindingPattern(parent.name)) {
@@ -316,7 +321,8 @@ export function parseScriptSetupRanges(
 				withDefaults = {
 					callExp: _getStartEnd(node),
 					exp: _getStartEnd(node.expression),
-					arg: arg ? _getStartEnd(arg) : undefined
+					arg: arg ? _getStartEnd(arg) : undefined,
+					argNode: arg
 				};
 			}
 			else if (vueCompilerOptions.macros.defineEmits.includes(callText)) {
@@ -345,12 +351,7 @@ export function parseScriptSetupRanges(
 					statement: getStatementRange(ts, parents, node, ast)
 				};
 				if (ts.isVariableDeclaration(parent)) {
-					if (ts.isIdentifier(parent.name)) {
-						defineSlots.name = _getNodeText(parent.name);
-					}
-					else {
-						defineSlots.isObjectBindingPattern = ts.isObjectBindingPattern(parent.name);
-					}
+					defineSlots.name = _getNodeText(parent.name);
 				}
 			}
 			else if (vueCompilerOptions.macros.defineExpose.includes(callText)) {