+++ packages/typescript-plugin/lib/server.ts
@@ -3,7 +3,11 @@ import * as fs from 'node:fs';
 import * as net from 'node:net';
 import type * as ts from 'typescript';
 import { collectExtractProps } from './requests/collectExtractProps';
-import { type ComponentPropInfo, getComponentDirectives, getComponentEvents, getComponentNames, getComponentProps, getElementAttrs } from './requests/componentInfos';
+import { getComponentDirectives } from './requests/getComponentDirectives';
+import { getComponentEvents } from './requests/getComponentEvents';
+import { getComponentNames } from './requests/getComponentNames';
+import { type ComponentPropInfo, getComponentProps } from './requests/getComponentProps';
+import { getElementAttrs } from './requests/getElementAttrs';
 import { getImportPathForFile } from './requests/getImportPathForFile';
 import { getPropertiesAtLocation } from './requests/getPropertiesAtLocation';
 import { getQuickInfoAtPosition } from './requests/getQuickInfoAtPosition';
@@ -102,7 +106,7 @@ export async function startNamedPipeServer(
 		connection.on('error', err => console.error('[Vue Named Pipe Server]', err.message));
 
 		for (const [fileName, [componentNames, componentProps]] of currentData) {
-			notify(connection, 'componentNamesUpdated', fileName, Object.keys(componentNames));
+			notify(connection, 'componentNamesUpdated', fileName, componentNames);
 
 			for (const [name, props] of Object.entries(componentProps)) {
 				notify(connection, 'componentPropsUpdated', fileName, [name, props]);