+++ test-workspace/component-meta/generic/component.vue
@@ -2,6 +2,6 @@
 defineProps<{ foo: number }>();
 defineEmits<{ (e: 'bar', data: number): void }>();
 defineExpose({ baz: {} as number });
-defineSlots<{ default?(data: { foo: number }): any }>();
+defineSlots<{ default?: (props: { foo: number }) => any }>();
 defineModel<string>('title');
 </script>