+++ test-workspace/tsc/passedFixtures/vue3/#4979/main.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 defineSlots<{
-	default(props: { foo: string }): any;
+	default: (props: { foo: string }) => any;
 }>();
 </script>